+++ package.json
@@ -10,6 +10,7 @@
   "dependencies": {
     "@headlessui/react": "^1.1.1",
     "abi-decoder": "^2.4.0",
+    "clsx": "^1.1.1",
     "csvtojson": "^2.0.10",
     "dotenv": "^10.0.0",
     "next": "^10.2.3",
