+++ styles/Home.module.css
@@ -2,42 +2,6 @@
   @apply text-5xl mt-14 mb-7 text-center font-semibold;
 }
 
-.search {
-  @apply inline-flex self-center bg-gray-200 items-center mb-7;
-  padding: 0 0 0 1.3rem;
-  border-radius: 2rem;
-}
-
-.search span {
-  @apply hidden sm:inline;
-}
-
-.search input {
-  margin: 0 0 0 1rem;
-  padding: .5rem;
-}
-
-.search button {
-  text-align: center;
-  padding: .6rem 1.3rem .5rem .75rem;
-  font-family: 'Nunito Sans', sans-serif;
-  color: white;
-  border-radius: 0 2rem 2rem 0;
-  align-self: center;
-}
-
-.search button:hover {
-  color: white;
-  box-shadow: 4px 4px 16px #f1f1f1, -4px -4px 16px #e6e6e6;
-  background: linear-gradient(-60deg, #e44e4e, #fa5549, #ffd035, #fa9049, #ff5151);
-  background-size: 400% 400%;
-  animation: Gradient 9s ease infinite;
-  -moz-animation: Gradient 9s ease infinite;
-  -webkit-animation: Gradient 9s ease infinite;
-  transition: 0.1s;
-  outline: none;
-}
-
 .emoji {
   @apply hidden sm:inline;
 }
@@ -91,97 +55,4 @@
 
 .tipjar-address {
   @apply mx-1 px-1 border-solid border rounded border-pink-500 text-pink-500;
-}
-
-.title a {
-  color: #0070f3;
-  text-decoration: none;
-}
-
-.title a:hover,
-.title a:focus,
-.title a:active {
-  text-decoration: underline;
-}
-
-.title {
-  margin: 0;
-  line-height: 1.15;
-  font-size: 4rem;
-}
-
-.title,
-.description {
-  text-align: center;
-}
-
-.description {
-  line-height: 1.5;
-  font-size: 1.5rem;
-}
-
-.code {
-  background: #fafafa;
-  border-radius: 5px;
-  padding: 0.75rem;
-  font-size: 1.1rem;
-  font-family: Menlo, Monaco, Lucida Console, Liberation Mono, DejaVu Sans Mono,
-    Bitstream Vera Sans Mono, Courier New, monospace;
-}
-
-.grid {
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  flex-wrap: wrap;
-  max-width: 800px;
-  margin-top: 3rem;
-}
-
-.card {
-  margin: 1rem;
-  padding: 1.5rem;
-  text-align: left;
-  color: inherit;
-  text-decoration: none;
-  border: 1px solid #eaeaea;
-  border-radius: 10px;
-  transition: color 0.15s ease, border-color 0.15s ease;
-  width: 45%;
-}
-
-.card:hover,
-.card:focus,
-.card:active {
-  color: #0070f3;
-  border-color: #0070f3;
-}
-
-.card h2 {
-  margin: 0 0 1rem 0;
-  font-size: 1.5rem;
-}
-
-.card p {
-  margin: 0;
-  font-size: 1.25rem;
-  line-height: 1.5;
-}
-
-.logo {
-  height: 1em;
-  margin-left: 0.5rem;
-}
-
-@media (max-width: 600px) {
-  .grid {
-    width: 100%;
-    flex-direction: column;
-  }
-}
-
-.search input[type=number]::-webkit-inner-spin-button,
-.search input[type=number]::-webkit-outer-spin-button {
-  -webkit-appearance: none;
-  margin: 0;
 }
